site-factory/requirements.txt
# Site Factory Engine 2.0
pillow>=10.0.0
requests>=2.31.0
supabase>=2.0.0
pyyaml>=6.0
